+++ tests/encoders/controllers/test_infer_body.py
@@ -92,3 +92,22 @@ def test_infer_body_with_dependency_without_provides(test_client_factory):
 
         assert response.status_code == 200
         assert response.json() == {"name": "lilya", "age": 10, "sku": "test", "x": "app_value"}
+
+
+class FormController(Controller):
+    async def post(self, user: User, item: Item):
+        return {**user.model_dump(), "sku": item.sku}
+
+
+def test_infer_body_from_form(test_client_factory):
+    data = {"user": '{"name": "lilya", "age": 10}', "item": '{"sku": "test"}'}
+
+    with create_client(
+        routes=[Path("/infer-form", handler=FormController)],
+        settings_module=EncoderSettings,
+    ) as client:
+        # send as form
+        response = client.post("/infer-form", data=data)
+
+        assert response.status_code == 200
+        assert response.json() == {"name": "lilya", "age": 10, "sku": "test"}